backend/requirements.txt
# =============================================================================
# NESTSECURE - Dependencias del Backend
# =============================================================================
# Versión: 1.0.0
# Python: 3.11+
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
fastapi==0.109.2
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# -----------------------------------------------------------------------------
# Base de Datos
# -----------------------------------------------------------------------------
sqlalchemy>=2.0.30       # ORM async (2.0.30+ para Python 3.13)
psycopg[binary,pool]>=3.1.0  # Driver async/sync para PostgreSQL (psycopg3)
alembic>=1.13.1          # Migraciones

# -----------------------------------------------------------------------------
# Cache y Message Broker
# -----------------------------------------------------------------------------
redis[hiredis]==5.0.1
celery[redis]==5.3.6

# -----------------------------------------------------------------------------
# Autenticación y Seguridad
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# -----------------------------------------------------------------------------
# Validación y Utilidades
# -----------------------------------------------------------------------------
email-validator==2.1.0
python-dateutil==2.8.2
pytz==2024.1

# -----------------------------------------------------------------------------
# HTTP Client
# -----------------------------------------------------------------------------
httpx==0.26.0
aiohttp==3.9.3

# -----------------------------------------------------------------------------
# Scanners y Parsing
# -----------------------------------------------------------------------------
python-nmap==0.7.1       # Wrapper para Nmap
defusedxml==0.7.1        # XML parsing seguro
lxml==5.1.0              # Parsing XML/HTML

# -----------------------------------------------------------------------------
# CVE y Vulnerabilidades
# -----------------------------------------------------------------------------
nvdlib==0.7.6            # NVD API client

# -----------------------------------------------------------------------------
# Reportes
# -----------------------------------------------------------------------------
jinja2==3.1.3
weasyprint==61.0
matplotlib==3.8.2

# -----------------------------------------------------------------------------
# Logging y Monitoreo
# -----------------------------------------------------------------------------
structlog==24.1.0
python-json-logger==2.0.7

# -----------------------------------------------------------------------------
# Utilidades Generales
# -----------------------------------------------------------------------------
tenacity==8.2.3          # Retry logic
orjson==3.9.13           # JSON rápido

# -----------------------------------------------------------------------------
# CLI
# -----------------------------------------------------------------------------
typer[all]==0.9.0
rich==13.7.0
